Seperate Nested File Traits of Row Map
