+++ app/Core/Download/RowMap.php
@@ -38,46 +38,3 @@ trait RowMap {
         }
     }
 }
-
-trait DeriveOccurrenceReference {
-    private static function derive_references($row) {
-        if (array_key_exists('occid', $row)) {
-            $path = '/occurrence/' . $row['occid'];
-            try {
-                return url($path);
-            } catch (\Throwable $th) {
-                return $path;
-            }
-        } else {
-            return;
-        }
-    }
-}
-
-trait DeriveCombineOccurrenceRecordID {
-    private static function derive_combine_occurrence_record_id($row) {
-        if (array_key_exists('recordID', $row) && array_key_exists('occurrenceID', $row)) {
-            $hasRecordID = $row['recordID'];
-            $hasOccurrenceID = $row['occurrenceID'];
-
-            if (! $hasRecordID && $hasOccurrenceID) {
-                return $row['occurrenceID'];
-            } elseif ($hasRecordID && ! $hasOccurrenceID) {
-                return $row['recordID'];
-            }
-        }
-    }
-}
-
-trait DeriveTaxonRank {
-    private static function derive_taxon_rank($row) {
-        //TODO (Logan) get better logic for this
-        if (array_key_exists('infraspecificEpithet', $row) && $row['infraspecificEpithet']) {
-            return 'Subspecies';
-        } elseif (array_key_exists('specificEpithet', $row) && $row['specificEpithet']) {
-            return 'Species';
-        } elseif (array_key_exists('genus', $row) && $row['genus']) {
-            return 'Genus';
-        }
-    }
-}
